+++ cli/app_test.go
@@ -25,7 +25,6 @@
 package cli
 
 import (
-	"context"
 	"testing"
 	"time"
 
@@ -37,20 +36,15 @@ import (
 	commonpb "go.temporal.io/api/common/v1"
 	enumspb "go.temporal.io/api/enums/v1"
 	historypb "go.temporal.io/api/history/v1"
-	namespacepb "go.temporal.io/api/namespace/v1"
 	"go.temporal.io/api/operatorservice/v1"
 	"go.temporal.io/api/operatorservicemock/v1"
-	replicationpb "go.temporal.io/api/replication/v1"
-	"go.temporal.io/api/serviceerror"
 	taskqueuepb "go.temporal.io/api/taskqueue/v1"
-	workflowpb "go.temporal.io/api/workflow/v1"
 	"go.temporal.io/api/workflowservice/v1"
 	"go.temporal.io/api/workflowservicemock/v1"
 	sdkclient "go.temporal.io/sdk/client"
 	sdkmocks "go.temporal.io/sdk/mocks"
 	healthpb "google.golang.org/grpc/health/grpc_health_v1"
 
-	"go.temporal.io/server/common/payloads"
 	"go.temporal.io/server/common/primitives/timestamp"
 )
 
@@ -126,362 +120,9 @@ func (s *cliAppSuite) TestAppCommands() {
 	}
 }
 
-func (s *cliAppSuite) TestNamespaceRegister_LocalNamespace() {
-	s.frontendClient.EXPECT().RegisterNamespace(gomock.Any(), gomock.Any()).Return(nil, nil)
-	err := s.app.Run([]string{"", "namespace", "register", "--global", "false", cliTestNamespace})
-	s.NoError(err)
-}
-
-func (s *cliAppSuite) TestNamespaceRegister_GlobalNamespace() {
-	s.frontendClient.EXPECT().RegisterNamespace(gomock.Any(), gomock.Any()).Return(nil, nil)
-	err := s.app.Run([]string{"", "namespace", "register", "--global", "true", cliTestNamespace})
-	s.NoError(err)
-}
-
-func (s *cliAppSuite) TestNamespaceRegister_NamespaceExist() {
-	s.frontendClient.EXPECT().RegisterNamespace(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewNamespaceAlreadyExists(""))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "register", "--global", "true", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestNamespaceRegister_Failed() {
-	s.frontendClient.EXPECT().RegisterNamespace(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewInvalidArgument("faked error"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "register", "--global", "true", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-var describeNamespaceResponseServer = &workflowservice.DescribeNamespaceResponse{
-	NamespaceInfo: &namespacepb.NamespaceInfo{
-		Name:        "test-namespace",
-		Description: "a test namespace",
-		OwnerEmail:  "test@uber.com",
-	},
-	Config: &namespacepb.NamespaceConfig{
-		WorkflowExecutionRetentionTtl: timestamp.DurationPtr(3 * time.Hour * 24),
-	},
-	ReplicationConfig: &replicationpb.NamespaceReplicationConfig{
-		ActiveClusterName: "active",
-		Clusters: []*replicationpb.ClusterReplicationConfig{
-			{
-				ClusterName: "active",
-			},
-			{
-				ClusterName: "standby",
-			},
-		},
-	},
-}
-
-func (s *cliAppSuite) TestNamespaceUpdate() {
-	resp := describeNamespaceResponseServer
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), gomock.Any()).Return(resp, nil).Times(2)
-	s.frontendClient.EXPECT().UpdateNamespace(gomock.Any(), gomock.Any()).Return(nil, nil).Times(2)
-	err := s.app.Run([]string{"", "namespace", "update", cliTestNamespace})
-	s.Nil(err)
-	err = s.app.Run([]string{"", "namespace", "update", "--description", "another desc", "--email", "another@uber.com", "--retention", "1", cliTestNamespace})
-	s.Nil(err)
-}
-
-func (s *cliAppSuite) TestNamespaceUpdate_NamespaceNotExist() {
-	resp := describeNamespaceResponseServer
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), gomock.Any()).Return(resp, nil)
-	s.frontendClient.EXPECT().UpdateNamespace(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewNamespaceNotFound("missing-namespace"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "update", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestNamespaceUpdate_ActiveClusterFlagNotSet_NamespaceNotExist() {
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewNamespaceNotFound("missing-namespace"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "update", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestNamespaceUpdate_Failed() {
-	resp := describeNamespaceResponseServer
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), gomock.Any()).Return(resp, nil)
-	s.frontendClient.EXPECT().UpdateNamespace(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewInvalidArgument("faked error"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "update", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestNamespaceDescribe() {
-	resp := describeNamespaceResponseServer
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), &workflowservice.DescribeNamespaceRequest{Namespace: cliTestNamespace, Id: ""}).Return(resp, nil)
-	err := s.app.Run([]string{"", "namespace", "describe", cliTestNamespace})
-	s.Nil(err)
-}
-
-func (s *cliAppSuite) TestNamespaceDescribe_ById() {
-	resp := describeNamespaceResponseServer
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), &workflowservice.DescribeNamespaceRequest{Namespace: "", Id: "nid"}).Return(resp, nil)
-	err := s.app.Run([]string{"", "namespace", "describe", "--namespace-id", "nid"})
-	s.Nil(err)
-}
-
-func (s *cliAppSuite) TestNamespaceDescribe_NamespaceNotExist() {
-	resp := describeNamespaceResponseServer
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), gomock.Any()).Return(resp, serviceerror.NewNamespaceNotFound("missing-namespace"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "describe", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestNamespaceDescribe_Failed() {
-	resp := describeNamespaceResponseServer
-	s.frontendClient.EXPECT().DescribeNamespace(gomock.Any(), gomock.Any()).Return(resp, serviceerror.NewInvalidArgument("faked error"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "describe", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestNamespaceDelete() {
-	s.operatorClient.EXPECT().DeleteNamespace(gomock.Any(), &operatorservice.DeleteNamespaceRequest{Namespace: cliTestNamespace}).Return(&operatorservice.DeleteNamespaceResponse{}, nil)
-	err := s.app.Run([]string{"", "namespace", "delete", "--yes", cliTestNamespace})
-	s.Nil(err)
-}
-
-func (s *cliAppSuite) TestNamespaceDelete_NamespaceNotExist() {
-	s.operatorClient.EXPECT().DeleteNamespace(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewNamespaceNotFound("missing-namespace"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "delete", "--yes", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestNamespaceDelete_Failed() {
-	s.operatorClient.EXPECT().DeleteNamespace(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewInvalidArgument("faked error"))
-	errorCode := s.RunWithExitCode([]string{"", "namespace", "delete", "--yes", cliTestNamespace})
-	s.Equal(1, errorCode)
-}
-
 var (
 	eventType = enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_STARTED
 )
-
-func (s *cliAppSuite) TestShowHistory() {
-	s.sdkClient.On("GetWorkflowHistory", mock.Anything, "wid", "", mock.Anything, mock.Anything).Return(historyEventIterator()).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "show", "--workflow-id", "wid"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestShowHistoryWithFollow() {
-	s.sdkClient.On("GetWorkflowHistory", mock.Anything, "wid", "", mock.Anything, mock.Anything).Return(historyEventIterator()).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "show", "--workflow-id", "wid", "--follow"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-
-	s.sdkClient.On("GetWorkflowHistory", mock.Anything, "wid", "", mock.Anything, mock.Anything).Return(historyEventIterator()).Once()
-	err = s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "show", "--workflow-id", "wid", "--fields", "long", "--follow"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestStartWorkflow() {
-	s.sdkClient.On("ExecuteWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(workflowRun(), nil)
-
-	// start with wid
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "start", "--task-queue", "testTaskQueue", "--type", "testWorkflowType", "--execution-timeout", "60", "--run-timeout", "60", "--workflow-id", "wid", "--workflow-id-reuse-policy", "Unspecified"})
-	s.Nil(err)
-	s.sdkClient.AssertNotCalled(s.T(), "GetWorkflowHistory")
-	s.sdkClient.AssertExpectations(s.T())
-
-	hasNoSearchAttributes := mock.MatchedBy(func(options sdkclient.StartWorkflowOptions) bool {
-		return len(options.SearchAttributes) == 0
-	})
-	s.sdkClient.AssertCalled(s.T(), "ExecuteWorkflow", mock.Anything, hasNoSearchAttributes, mock.Anything, mock.Anything)
-
-	s.sdkClient.On("ExecuteWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(workflowRun(), nil)
-	// start without wid
-	err = s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "start", "--task-queue", "testTaskQueue", "--type", "testWorkflowType", "--execution-timeout", "60", "--run-timeout", "60", "--workflow-id-reuse-policy", "Unspecified"})
-	s.Nil(err)
-	s.sdkClient.AssertNotCalled(s.T(), "GetWorkflowHistory")
-	s.sdkClient.AssertExpectations(s.T())
-	s.sdkClient.AssertCalled(s.T(), "ExecuteWorkflow", mock.Anything, hasNoSearchAttributes, mock.Anything, mock.Anything)
-}
-
-func (s *cliAppSuite) TestStartWorkflow_SearchAttributes() {
-	s.sdkClient.On("ExecuteWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(workflowRun(), nil)
-	// start with basic search attributes
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "start", "--task-queue", "testTaskQueue", "--type", "testWorkflowType",
-		"--search-attribute-key", "k1", "--search-attribute-value", "\"v1\"", "--search-attribute-key", "k2", "--search-attribute-value", "\"v2\""})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-
-	hasCorrectSearchAttributes := mock.MatchedBy(func(options sdkclient.StartWorkflowOptions) bool {
-		return len(options.SearchAttributes) == 2 &&
-			options.SearchAttributes["k1"] == "v1" &&
-			options.SearchAttributes["k2"] == "v2"
-	})
-	s.sdkClient.AssertCalled(s.T(), "ExecuteWorkflow", mock.Anything, hasCorrectSearchAttributes, mock.Anything, mock.Anything)
-
-	s.sdkClient.On("ExecuteWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(workflowRun(), nil)
-
-	// TODO: test json search attributes once we know how to they'll be specified (--search-attribute-json?)
-}
-
-func (s *cliAppSuite) TestStartWorkflow_Failed() {
-	s.sdkClient.On("ExecuteWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(workflowRun(), serviceerror.NewInvalidArgument("fake error"))
-
-	// start with wid
-	errorCode := s.RunWithExitCode([]string{"", "--namespace", cliTestNamespace, "workflow", "start", "--task-queue", "testTaskQueue", "--type", "testWorkflowType", "--execution-timeout", "60", "--run-timeout", "60", "--workflow-id", "wid"})
-	s.Equal(1, errorCode)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestExecuteWorkflow() {
-	s.sdkClient.On("ExecuteWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(workflowRun(), nil)
-	s.sdkClient.On("GetWorkflowHistory", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(historyEventIterator()).Once()
-
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "execute", "--task-queue", "testTaskQueue", "--type", "testWorkflowType"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestTerminateWorkflow() {
-	s.sdkClient.On("TerminateWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(nil).Once()
-
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "terminate", "--workflow-id", "wid"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestTerminateWorkflow_Failed() {
-	s.sdkClient.On("TerminateWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(serviceerror.NewInvalidArgument("faked error")).Once()
-	errorCode := s.RunWithExitCode([]string{"", "--namespace", cliTestNamespace, "workflow", "terminate", "--workflow-id", "wid"})
-	s.Equal(1, errorCode)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestCancelWorkflow() {
-	s.sdkClient.On("CancelWorkflow", mock.Anything, mock.Anything, mock.Anything).Return(nil).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "cancel", "--workflow-id", "wid"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestCancelWorkflow_Failed() {
-	s.sdkClient.On("CancelWorkflow", mock.Anything, mock.Anything, mock.Anything).Return(serviceerror.NewInvalidArgument("faked error")).Once()
-	errorCode := s.RunWithExitCode([]string{"", "--namespace", cliTestNamespace, "workflow", "cancel", "--workflow-id", "wid"})
-	s.Equal(1, errorCode)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestSignalWorkflow() {
-	s.frontendClient.EXPECT().SignalWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil, nil)
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "signal", "--name", "signal-name", "--workflow-id", "wid"})
-	s.Nil(err)
-}
-
-func (s *cliAppSuite) TestSignalWorkflow_Failed() {
-	s.frontendClient.EXPECT().SignalWorkflowExecution(gomock.Any(), gomock.Any()).Return(nil, serviceerror.NewInvalidArgument("faked error"))
-	errorCode := s.RunWithExitCode([]string{"", "--namespace", cliTestNamespace, "workflow", "signal", "--name", "signal-name", "--workflow-id", "wid"})
-	s.Equal(1, errorCode)
-}
-
-func (s *cliAppSuite) TestQueryWorkflow() {
-	resp := &workflowservice.QueryWorkflowResponse{
-		QueryResult: payloads.EncodeString("query-result"),
-	}
-	s.frontendClient.EXPECT().QueryWorkflow(gomock.Any(), gomock.Any()).Return(resp, nil)
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "query", "--workflow-id", "wid", "--query-type", "query-type-test"})
-	s.Nil(err)
-}
-
-func (s *cliAppSuite) TestQueryWorkflowUsingStackTrace() {
-	resp := &workflowservice.QueryWorkflowResponse{
-		QueryResult: payloads.EncodeString("query-result"),
-	}
-	s.frontendClient.EXPECT().QueryWorkflow(gomock.Any(), gomock.Any()).Return(resp, nil)
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "stack", "--workflow-id", "wid"})
-	s.Nil(err)
-}
-
-func (s *cliAppSuite) TestQueryWorkflow_Failed() {
-	resp := &workflowservice.QueryWorkflowResponse{
-		QueryResult: payloads.EncodeString("query-result"),
-	}
-	s.frontendClient.EXPECT().QueryWorkflow(gomock.Any(), gomock.Any()).Return(resp, serviceerror.NewInvalidArgument("faked error"))
-	errorCode := s.RunWithExitCode([]string{"", "--namespace", cliTestNamespace, "workflow", "query", "--workflow-id", "wid", "--query-type", "query-type-test"})
-	s.Equal(1, errorCode)
-}
-
-var (
-	listWorkflowExecutionsResponse = &workflowservice.ListWorkflowExecutionsResponse{
-		Executions: []*workflowpb.WorkflowExecutionInfo{
-			{
-				Execution: &commonpb.WorkflowExecution{
-					WorkflowId: "test-list-open-workflow-id",
-					RunId:      uuid.New(),
-				},
-				Type: &commonpb.WorkflowType{
-					Name: "test-list-open-workflow-type",
-				},
-				StartTime:     timestamp.TimePtr(time.Now().UTC()),
-				CloseTime:     timestamp.TimePtr(time.Now().UTC().Add(time.Hour)),
-				HistoryLength: 12,
-			},
-			{
-				Execution: &commonpb.WorkflowExecution{
-					WorkflowId: "test-list-workflow-id",
-					RunId:      uuid.New(),
-				},
-				Type: &commonpb.WorkflowType{
-					Name: "test-list-workflow-type",
-				},
-				StartTime:     timestamp.TimePtr(time.Now().UTC()),
-				CloseTime:     timestamp.TimePtr(time.Now().UTC().Add(time.Hour)),
-				Status:        enumspb.WORKFLOW_EXECUTION_STATUS_COMPLETED,
-				HistoryLength: 12,
-			},
-		},
-	}
-)
-
-func (s *cliAppSuite) TestListWorkflow() {
-	s.sdkClient.On("ListWorkflow", mock.Anything, mock.Anything).Return(listWorkflowExecutionsResponse, nil).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "list"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestListWorkflow_DeadlineExceeded() {
-	s.sdkClient.On("ListWorkflow", mock.Anything, mock.Anything).Return(nil, context.DeadlineExceeded).Once()
-	s.sdkClient.On("ListWorkflow", mock.Anything, mock.Anything).Return(listWorkflowExecutionsResponse, nil).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "list"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestListWorkflow_Open_WithQuery() {
-	s.sdkClient.On("ListWorkflow", mock.Anything, mock.Anything).Return(listWorkflowExecutionsResponse, nil).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "list", "--query", "ExecutionStatus='Running'"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestListArchivedWorkflow() {
-	s.sdkClient.On("ListArchivedWorkflow", mock.Anything, mock.Anything).Return(&workflowservice.ListArchivedWorkflowExecutionsResponse{}, nil).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "list", "--archived", "--query", "some query string"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestCountWorkflow() {
-	s.sdkClient.On("CountWorkflow", mock.Anything, mock.Anything).Return(&workflowservice.CountWorkflowExecutionsResponse{}, nil).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "count"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-
-	s.sdkClient.On("CountWorkflow", mock.Anything, mock.Anything).Return(&workflowservice.CountWorkflowExecutionsResponse{}, nil).Once()
-	err = s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "count", "--query", "'CloseTime = missing'"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
-
-func (s *cliAppSuite) TestCountWorkflowDeadlineExceeded() {
-	s.sdkClient.On("CountWorkflow", mock.Anything, mock.Anything).Return(nil, context.DeadlineExceeded).Once()
-	s.sdkClient.On("CountWorkflow", mock.Anything, mock.Anything).Return(&workflowservice.CountWorkflowExecutionsResponse{}, nil).Once()
-	err := s.app.Run([]string{"", "--namespace", cliTestNamespace, "workflow", "count", "--query", "'CloseTime = missing'"})
-	s.Nil(err)
-	s.sdkClient.AssertExpectations(s.T())
-}
 
 var describeTaskQueueResponse = &workflowservice.DescribeTaskQueueResponse{
 	Pollers: []*taskqueuepb.PollerInfo{
